Incorporated feedback from code review: Converted unit tests to XUnit, Removed hard coded reference path
